requirements.txt
whois
requests
tldextract
dnspython
pandas
bs4
python-dotenv
numpy
scikit-learn
setuptools
pymongo
certifi
pyaml
mlflow
dagshub
fastapi
uvicorn
jinja2
python-multipart


##-e .